Сделал определение местоположения. Без использования Google Maps.

+++ app/src/main/java/ru/gressor/weatherapp/data_types/HistoryStorage.java
@@ -61,8 +61,8 @@ public class HistoryStorage implements Parcelable {
 
         for (int i = 0; i < cities.size(); i++) {
             City city = cities.get(i);
-            PositionPoint positionPoint = new PositionPoint(city.getCityName(), null);
-            positionPoint.setCoord(city.getCoord());
+            PositionPoint positionPoint =
+                    new PositionPoint(city.getCityName(), null, city.getCoord());
 
             WeatherState weatherState = new WeatherState(
                     new ActualWeather(city.getTemperature(), city.getActualAt()),
@@ -82,6 +82,8 @@ public class HistoryStorage implements Parcelable {
     }
 
     public void push(HistoryItem historyItem) {
+        if (historyItems.get(0).getPositionPoint().getTown() == null) historyItems.remove(0);
+
         boolean isFavorite = removeSimilar(historyItem);
         if (isFavorite) historyItem.setFavorite(true);
 
@@ -91,6 +93,7 @@ public class HistoryStorage implements Parcelable {
 
         LinkedList<HistoryItem> favorites = getSublistFirstExcluded(true);
         int favoritesCount = favorites.size() + (historyItem.isFavorite() ? 1 : 0);
+
         if (favoritesCount > MAX_FAVORITES) {
             favorites.get(favorites.size() - 1).setFavorite(false);
         }
@@ -127,27 +130,21 @@ public class HistoryStorage implements Parcelable {
         return result;
     }
 
-    public PositionPoint getCurrentPosition() {
-        return historyItems.get(0).getPositionPoint();
-    }
-
-    public WeatherState getCurrentWeather() {
-        return historyItems.get(0).getWeatherState();
-    }
-
     private boolean removeSimilar(HistoryItem historyItem) {
         return removeSimilar(historyItem.getPositionPoint().getTown());
     }
 
     private boolean removeSimilar(String townName) {
-        weatherDao.deleteCity(townName);
+        if (townName != null) {
+            weatherDao.deleteCity(townName);
 
-        for (int i = 0; i < historyItems.size(); i++) {
-            HistoryItem item = historyItems.get(i);
-            if (item.getPositionPoint().getTown().equals(townName)) {
-                boolean isFavorite = item.isFavorite();
-                historyItems.remove(item);
-                return isFavorite;
+            for (int i = 0; i < historyItems.size(); i++) {
+                HistoryItem item = historyItems.get(i);
+                if (item.getPositionPoint().getTown().equals(townName)) {
+                    boolean isFavorite = item.isFavorite();
+                    historyItems.remove(item);
+                    return isFavorite;
+                }
             }
         }
         return false;
@@ -172,6 +169,14 @@ public class HistoryStorage implements Parcelable {
             }
         }
         return result;
+    }
+
+    public PositionPoint getCurrentPosition() {
+        return historyItems.get(0).getPositionPoint();
+    }
+
+    public WeatherState getCurrentWeather() {
+        return historyItems.get(0).getWeatherState();
     }
 
     public static final Creator<HistoryStorage> CREATOR = new Creator<HistoryStorage>() {
